refactor: PostModal.jsx -> index.jsx
